Continue to work on odin-recipes project
Stop at Iterarion 4 where the project requests for the creating a total
of for recipe pages. Finished  chorizo.html which is the second recipe.
Continuing this, one more recipe is needed to finish the  project. Check
the TOP page to see if there are more todo for this project/lesson.